gjort så de använder sig av samma move i den abstrakta Entity
Tog bort allt med draw positioner

delade upp klassen som rör enemy i 2 delar, en som rör på den, en som säger om den är klar eller inte

+++ src/se/liu/antos931jakos322/towerdefence/entities/enemies/EnemyCreator.java
@@ -10,7 +10,7 @@ import java.awt.*;
 
 public class EnemyCreator extends Enemy
 {
-    public EnemyCreator(int health, int speed, Color color, double size, int damage) {
+    public EnemyCreator(int health, double speed, Color color, double size, int damage) {
 
         super(health, speed, color, size, damage);
     }
